♻️ Merge redundant skill_repository into skill_db.py
